latest version in lab 3.1

+++ CacheTable.h
@@ -2,7 +2,7 @@ class CacheTable{
 private:
 		char *data;
 		bool *valid;
-		int *tag;
+		long long int *tag;
 		bool *dirty;
 		int *lru;
 		int set_num;
@@ -13,7 +13,7 @@ public:
 	CacheTable(int set_num_,int line_num_,int offset_num_);
 	~CacheTable();
 	bool getValid(int set,int line);//get
-	int getTag(int set,int line);
+	long long int getTag(int set,int line);
 	char getData(int set,int line,int offset);
 	bool getDirty(int set,int line);
 	int getLRU(int set,int line);
@@ -23,7 +23,7 @@ public:
 
 
 	bool setValid(int set,int line,bool value);//set
-	bool setTag(int set,int line,int value);
+	bool setTag(int set,int line,long long int value);
 	bool setData(int set,int line,int offset,char value);
 	bool setSet_num(int value);
 	bool setLine_num(int value);
